integrated validate on mount prop.
