+++ src/forms/index.js
@@ -1,4 +1,3 @@
 export * from './finalFormPropTypes';
 export * from './renderFunctions';
-export { default as FinalFormSelect } from './finalFormSelect';
 export * from './finalFormComponent';
