+++ src/forms/tests/finalFormField.test.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { mount } from 'enzyme';
 import renderer from 'react-test-renderer';
-import { FinalFormField } from '../finalFormField';
+import { FinalFormField } from '../';
 import { fieldInputProp, fieldMetaProps } from './mocks';
 
 describe('Final form input component', () => {
